feature(pedalInfoValidation): Pedal populate to a single file
